Warning fixing

+++ src/Render/DeferredRenderer.cpp
@@ -9,7 +9,7 @@
 
 namespace Phoenix
 {
-	DeferredRenderer::DeferredRenderer(IRIDevice* renderDevice, size_t gBufferWidth, size_t gBufferHeight)
+	DeferredRenderer::DeferredRenderer(IRIDevice* renderDevice, uint32_t gBufferWidth, uint32_t gBufferHeight)
 	{
 		TextureDesc desc;
 		desc.pixelFormat = EPixelFormat::RGBA16F;
